Added channel and Mapping
